// server/src/enums/user.enum.ts
export enum UserRole {
    SUPERADMIN = 'superadmin',
    USER = 'user',
}

export enum UserLocation {
    MUMBAI = 'MUMBAI',
    BENGALURU = 'BENGALURU',
    OVERALL = 'OVERALL',
}

export enum UserDepartment {
    SALES = 'SALES',
    DATA = 'DATA',
    IT = 'IT',
    HR = 'HR',
    GENERAL = 'GENERAL',
}